+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 [Added]
 - Get IdP's friendly name from metadata
+- Get SP's friendly name from metadata
 
 ## [Unreleased]
 [Added]